requirements.txt
bert-extractive-summarizer==0.10.1
blis==0.4.1
boto3==1.26.83
botocore==1.29.83
catalogue==1.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cymem==2.0.7
Cython==0.29.33
filelock==3.9.0
huggingface-hub==0.12.1
idna==3.4
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
neuralcoref==4.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pathy==0.10.1
pkuseg==0.0.25
plac==1.1.3
preshed==3.0.8
pydantic==1.10.5
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
spacy==2.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==1.0.6
thinc==7.4.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==0.10.1
zh-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_lg-2.3.1/zh_core_web_lg-2.3.1.tar.gz
